requirements.txt
azure-ai-vision